fix the $when default parameter to use the correct syntax
If when is set to just '01:00' (as it was previously), then backups
will never run due to the time format being specified
incorrectly. Backupninja considers '01:00' to be the day of the week
and every time it wakes up to see if it is time to do a backup it
determines that it is currently not the day "01:00" and so it
shouldn't do a backup now. In otherwords, it will never do a backup at
this rate.

I've filed a bug for this issue, because it probably should at least
let you know that you've failed.

    https://labs.riseup.net/code/issues/show/640